임시저장

+++ src/main/java/com/xcurenet/emass/message/vo/emass/mongo/fields/DayVo_Mgo.java
@@ -2,6 +2,6 @@ package com.xcurenet.emass.message.vo.emass.mongo.fields;
 
 
 public class DayVo_Mgo {
-    private int	week;	//몇주차
+    private Long	week;	//몇주차
     private boolean	work;	//업무시간 여부
 }
